scheduler timesteps as float32 instead of float64
Having the scheduler timesteps default to flaot64 when all other variables are float32 is unecessary, but more importantly causes problems when running on Apple Silicon since float64 is not implemented in MPS but float32 is.

+++ src/diffusers/schedulers/scheduling_lms_discrete.py
@@ -205,7 +205,7 @@ class LMSDiscreteScheduler(SchedulerMixin, ConfigMixin):
         """
         self.num_inference_steps = num_inference_steps
 
-        timesteps = np.linspace(0, self.config.num_train_timesteps - 1, num_inference_steps, dtype=float)[::-1].copy()
+        timesteps = np.linspace(0, self.config.num_train_timesteps - 1, num_inference_steps, dtype=np.float32)[::-1].copy()
 
         sigmas = np.array(((1 - self.alphas_cumprod) / self.alphas_cumprod) ** 0.5)
         log_sigmas = np.log(sigmas)
